General repo improvements

+++ contracts/contracts/coordination/StakeInfo.sol
@@ -7,11 +7,10 @@ import "./IUpdatableStakeInfo.sol";
 import "../../threshold/IAccessControlApplication.sol";
 
 /**
-* @title StakeInfo
-* @notice StakeInfo
-*/
+ * @title StakeInfo
+ * @notice StakeInfo
+ */
 contract StakeInfo is AccessControl, IUpdatableStakeInfo, IAccessControlApplication {
-
     bytes32 public constant UPDATE_ROLE = keccak256("UPDATE_ROLE");
 
     struct Stake {
@@ -20,8 +19,8 @@ contract StakeInfo is AccessControl, IUpdatableStakeInfo, IAccessControlApplicat
         // TODO: what about undelegations etc?
     }
 
-    constructor(address[] memory updaters){
-        for(uint i = 0; i < updaters.length; i++){
+    constructor(address[] memory updaters) {
+        for (uint256 i = 0; i < updaters.length; i++) {
             _grantRole(UPDATE_ROLE, updaters[i]);
         }
     }
@@ -29,19 +28,25 @@ contract StakeInfo is AccessControl, IUpdatableStakeInfo, IAccessControlApplicat
     mapping(address => Stake) public stakes;
     mapping(address => address) private operatorToProvider;
 
-    function stakingProviderFromOperator(address _operator) external view returns (address){
+    function stakingProviderFromOperator(address _operator) external view returns (address) {
         return operatorToProvider[_operator];
     }
 
-    function authorizedStake(address _stakingProvider) external view returns (uint96){
+    function authorizedStake(address _stakingProvider) external view returns (uint96) {
         return stakes[_stakingProvider].amount;
     }
 
-    function updateOperator(address stakingProvider, address operator) external override onlyRole(UPDATE_ROLE) {
+    function updateOperator(
+        address stakingProvider,
+        address operator
+    ) external override onlyRole(UPDATE_ROLE) {
         _updateOperator(stakingProvider, operator);
     }
 
-    function updateAmount(address stakingProvider, uint96 amount) external override onlyRole(UPDATE_ROLE) {
+    function updateAmount(
+        address stakingProvider,
+        uint96 amount
+    ) external override onlyRole(UPDATE_ROLE) {
         _updateAmount(stakingProvider, amount);
     }
 
@@ -49,7 +54,7 @@ contract StakeInfo is AccessControl, IUpdatableStakeInfo, IAccessControlApplicat
         Stake storage stake = stakes[stakingProvider];
         address oldOperator = stake.operator;
 
-        if(operator != oldOperator){
+        if (operator != oldOperator) {
             stake.operator = operator;
             // Update operator to provider mapping
             operatorToProvider[oldOperator] = address(0);
@@ -62,8 +67,8 @@ contract StakeInfo is AccessControl, IUpdatableStakeInfo, IAccessControlApplicat
     function _updateAmount(address stakingProvider, uint96 amount) internal {
         Stake storage stake = stakes[stakingProvider];
         uint256 oldAmount = stake.amount;
-        
-        if(amount != oldAmount){
+
+        if (amount != oldAmount) {
             stake.amount = amount;
             emit UpdatedStakeAmount(stakingProvider, amount);
         }
@@ -71,16 +76,16 @@ contract StakeInfo is AccessControl, IUpdatableStakeInfo, IAccessControlApplicat
 
     function batchUpdate(bytes32[] calldata updateInfo) external override onlyRole(UPDATE_ROLE) {
         require(updateInfo.length % 2 == 0, "bad length");
-        for(uint i = 0; i < updateInfo.length; i += 2){
+        for (uint256 i = 0; i < updateInfo.length; i += 2) {
             bytes32 word0 = updateInfo[i];
             bytes32 word1 = updateInfo[i + 1];
-            
+
             address provider = address(bytes20(word0));
             uint96 amount = uint96(uint256(word0));
             address operator = address(bytes20(word1));
 
             _updateOperator(provider, operator);
-            _updateAmount(provider, amount); 
+            _updateAmount(provider, amount);
         }
     }
 }
